Use mock GC implementation for MemoryMonitorTests
